+++ CHANGELOG.md
@@ -48,3 +48,4 @@ Version 0.6.0 (2018-05-30)
 * Raise minimum Rust version to 1.26
 * Add `generate_unique_from(sudoku)`
 * Add `shuffle()` which transforms a sudoku into a random equivalent sudoku.
+* Add `n_clues()` for the number of filled cells